added imgs (subjects, levels)

+++ controller/subject.js
@@ -1,6 +1,8 @@
 const asyncHandler = require("express-async-handler");
 const Subject = require("../modules/Subject");
 const ApiSuccess = require("../utils/apiSuccess");
+const uploadController = require("./upload");
+const ApiError = require("../utils/apiError");
 
 const getAll = asyncHandler(async (req, res) => {
   const subjects = await Subject.find().populate("img");
@@ -21,12 +23,21 @@ const update = asyncHandler(async (req, res) => {
     }
   );
   return res.json(new ApiSuccess("Updated subject successfully", subject));
-});
-const remove = asyncHandler(async (req, res) => {
-  const subject = await Subject.findByIdAndDelete(req.query._id);
-  return res.json(new ApiSuccess("Deleted subject successfully", subject));
-});
+}); 
+const remove = asyncHandler(async (req, res, next) => {
+  const subject = await Subject.findById(req.query._id).populate("img");
+  if (!subject) {
+    return next(new ApiError("Subject not found", 404));
+  }
+  if (subject.img) {
+    req.body.owner = subject.img.owner;
+    req.body.fileName = subject.img.fileName;
 
+    await uploadController.remove(req, res, next);
+  }
+  await Subject.findByIdAndDelete(req.query._id);
+  new ApiSuccess("Deleted subject and related file successfully", subject);
+});
 module.exports = {
   getAll,
   create,
